example_project/requirements.txt
asgiref==3.6.0 ; python_version >= "3.11" and python_version < "4.0"
django-environ==0.9.0 ; python_version >= "3.11" and python_version < "4"
django-extensions==3.2.1 ; python_version >= "3.11" and python_version < "4.0"
django-filter==21.1 ; python_version >= "3.11" and python_version < "4.0"
django-model-utils==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
django==4.1.4 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
pytz==2022.7 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
